+++ src/cli.ts
@@ -3,7 +3,7 @@ import { CLI } from '@ideascol/cli-maker';
 import CommandCreate from './commands/create';
 import CommandDoctor from './commands/doctor';
 import CommandUpgrade from './commands/upgrade';
-import { PanelGroup, MenuGroup, CommandGroup, RpcGroup, StatusBarGroup, SubpanelGroup, TreeViewGroup, TestGroup, PublishGroup } from './commands/groups';
+import { PanelGroup, MenuGroup, CommandGroup, RpcGroup, StatusBarGroup, SubpanelGroup, TreeViewGroup, TestGroup, PublishGroup, HelperGroup } from './commands/groups';
 
 const cli = new CLI(
   'vsceasy',
@@ -34,6 +34,7 @@ cli.command(SubpanelGroup);
 cli.command(TreeViewGroup);
 cli.command(TestGroup);
 cli.command(PublishGroup);
+cli.command(HelperGroup);
 cli.command(CommandDoctor);
 cli.command(CommandUpgrade);